SW-1966 Add endpoint to update batch quantities (#685)
`PUT /api/v1/nursery/batches/{id}/quantities` will update the seedling quantities
of an existing batch.

This endpoint introduces a new pattern to the API: optimistic locking using
versions. This will help us manage data consistency when multiple users are
working with the same batch. The client must pass in a version number and the
version number must match the current version of the batch. If the versions
don't match, it means some other client updated the batch already, and the
server returns HTTP 412.

Example flow to demonstrate how it works:

1. Batch 123 has a version number of 9
2. Client A: `GET /api/v1/nursery/batches/123` (returns version 9)
3. Client B: `GET /api/v1/nursery/batches/123` (returns version 9)
4. Client A: `PUT /api/v1/nursery/batches/123/quantities` with version=9
   (succeeds and returns updated batch with version 10)
5. Client B: `PUT /api/v1/nursery/batches/123/quantities` with version=9
   (fails with HTTP 412)
6. Client B: `GET /api/v1/nursery/batches/123` (returns version 10)
7. Client B: `PUT /api/v1/nursery/batches/123/quantities` with version=10
   (succeeds and returns updated batch with version 11)