SW-1845 Revamp accession and batch number generation
Previously, accession numbers were globally unique, which was fine because we were
generating them on the server side. But now we allow users to upload CSV files
with their own arbitrary accession numbers, and we don't want one organization to
be able to make an accession number unavailable to another one.

Change the database constraint so that accession numbers are only required to be
unique per facility.

In addition, the fact that we encoded dates into accession numbers sometimes
causes user confusion: those dates _usually_ match up with several other accession
dates such as the collection date, but not always. Although the date selection was
consistent on the server side, users didn't have a good way of learning how it
worked and made assumptions which then broke some of the time.

To address that, we now allocate these identifiers as simple incrementing numbers,
tracked separately per organization, starting with 100000 so they'll stay a
consistent width on screen for all but the very largest of organizations.

Seed accession and seedling batch numbers are allocated from the same place; we
will never generate an accession number that we've generated as a batch number for
the same organization or vice versa. However, if the user generates their own
accession numbers and uploads them in a CSV file, we don't enforce that the
numbers have to be unique across the organization, just unique within a particular
seed bank.