SW-1845 Revamp accession and batch number generation
Previously, accession numbers were globally unique, which was fine because we were
generating them on the server side. But now we allow users to upload CSV files
with their own arbitrary accession numbers, and we don't want one organization to
be able to make an accession number unavailable to another one.

Change the database constraint so that accession numbers are only required to be
unique per facility, and change the number generation so that we generate numbers
per organization rather than globally. Note the difference between the constraint
(per-facility uniqueness) and the default (per-organization uniqueness) -- we
think the latter is a good idea, but if people are uploading their own inventory
spreadsheets, they might use the same numbering scheme at two seed banks and we
don't want to bomb out on the CSV upload.