Handle time zones; improve documentation; fix race condition
This started off as an edit to the documentation as prompted by code review
comments.

I got to the previous doc version's paragraph about there being a race
condition if two servers allocate identifiers on a new day at the same time and
realized I could fix that race condition, and simplify the application code,
with a tweak to the SQL query.

Then I was going to add a comment that the "current date" was always in UTC,
but realized adding the time zone as an optional argument would make the default
obvious in the code and would also come in handy down the road when we start
supporting non-UTC timezones.