SW-1986 Generate shorter accession/batch numbers
After getting feedback about the previous accession/batch numbering scheme,
we want to change two things: make the identifiers shorter and allow people
to visually differentiate between accession numbers and batch numbers.

This new format is 3 characters shorter and treats accessions and batches
as separate sequential number spaces: an accession number in the old format
looked like `20221012888` and in the new scheme it would be `22-1-888` where
the `1` indicates that it's an accession number.