SW-1979 Move jOOQ configuration to a subproject (#708)
Previously, our customization and configuration of jOOQ code generation (which
needs to be executed by Gradle as part of the build process, not by our
application code at runtime) lived in the `buildSrc` directory.

`buildSrc` has a problem: whenever it changes, the entire project is rebuilt.
Changing the jOOQ configuration would cause the MJML email templates to be
rebuilt even though they have no dependency relationship.

Move the jOOQ configuration to a separate subproject that produces a Gradle
plugin. Changes to the configuration will cause the plugin to be rebuilt, which
will trigger a fresh round of jOOQ code generation, but won't invalidate
unrelated parts of the project. Updating `Config.kt` won't cause MJML templates
to be rebuilt.