SW-1603 Detect collision of suggested species name
If species 1 and 2 exist, and species 2 has an incorrect name (misspelled or
synonym) that should be changed to the same name as species 1, return HTTP 409
(Conflict) with a descriptive error message rather than an internal server error
when the user tries to accept the suggested edit.

Eventually we'll most likely want to add logic to merge the two species, but
until then, we should at least show a meaningful error message.