SW-1939 Remove v1 accession payload classes
We were still using the v1 accession payload classes in unit tests. Update the
tests to use v2 model classes instead, and remove the payload classes.

To replace the previous "apply a PUT request payload to an accession model"
helper method, this change introduces a couple new helper methods to the
`AccessionStore` test suite that use a fluent API style.

Without the new API, migrating the tests to work directly with `AccessionModel`
objects would have caused them to require a bunch of noisy temporary variables
or awkward scope function calls.

In this initial PR, the new API is only used in test methods that needed to
change to work with v2 models.