SW-1939 Remove v1 accession endpoints
Now that the production web app is switched over to the v2 accessions API, we can
start to get rid of the v1 API and its implementation.

This initial change removes the v1 API endpoints that have been supplanted by v2
equivalents, and removes some code that was only referenced by the endpoints'
controller methods. Specifically, the following endpoints go away:

* `GET /api/v1/seedbank/accessions/{id}` (read accession)
* `POST /api/v1/seedbank/accessions` (create accession)
* `PUT /api/v1/seedbank/accessions/{id}` (update accession)

In addition, `POST /api/v1/seedbank/accessions/{id}/checkIn` now returns a v2
response payload instead of a v1 payload. Previously, you'd get back a v1-style
response even if you were checking in an accession you'd creeated with v2, and
you'd then have to immediately GET the v2 version.

Some of the payload classes are removed here as well, but not all of them; a
few are used in unit tests and will be removed when those tests are updated to
verify the v2 business logic instead of v1.