Use scheduled withdrawal date as observation time
In v1, we allowed withdrawals to be scheduled for future dates, and if the
accessions were weight-based, the user had to enter the (expected) remaining
weight ahead of time. These withdrawals had dates that were later than their
creation times, which was causing the v1->v2 conversion to calculate incorrect
(or even negative) remaining quantities on these accessions.

This fix is only relevant for v1->v2 conversion of existing accessions, but
without it, some of the accessions in the production environment can't be
correctly migrated to v2.