SW-1939 Remove accession v2->v1 conversion code (#717)
With the removal of the v1 API endpoints, there is no longer any way for a client
to ask the server to convert a v2-style accession to v1-style. Remove the
conversion code and its associated tests.

The v1->v2 conversion code is unaffected here; we'll continue to need it until all
existing accessions have been converted to v2.