SW-2019 Populate estimated quantities at creation time
In cases where an accession's quantity was known at creation time, e.g., when
importing from a CSV file, we weren't populating the estimated count/weight.
This was a holdover from the v1 world where quantity could only be set as a
processing step after the accession was initially created.

Start populating the missing values, and update the CSV importer test to
verify the desired behavior.