SW-1939 Remove summary calculations for scheduled withdrawals (#723)
In v1, we allowed withdrawals to be scheduled for the future, and we had summaries
of future withdrawals in the UI. In v2, we don't allow scheduling and there are no
summaries. Remove the code that calculatees the summary data.

This code was already dead; it was only referenced by tests.