SW-1939 Remove test cases for v1-only behavior
This is an initial coarse-grained pass at removing the tests that verify the
v1 accession behavior now that there's no longer a way to create v1 accessions.
It removes test methods whose only purpose is to test behavior that only
applies to v1, and thus would make no sense to try to update to use the v2 code.

Tests that cover both v1 and v2 behavior, or tests that use v1-specific APIs but
are testing functionality that has a v2 equivalent, are not modified here. The
intent is to reduce the surface area of the test suite in a safe, easy-to-review
way so it becomes faster to update the remaining tests.