Add missing descriptions to OpenAPI schema fields (#725)
Our OpenAPI customization class has code to work around a limitation of the
Swagger core library's annotation processing that causes it to throw away payload
field descriptions if the fields are references to other payload classes.
Previously, the workaround required keeping an explicit list of all the payload
classes that needed to be fixed up. The list was, inevitably, out of date and
missing several classes that ought to have been included.

Switch to a classpath-scanning approach where we probe the code base using
reflection and find all the payload classes. This is slower but it's a one-time
operation that only happens the first time someone requests the OpenAPI schema
from a server instance.

In addition, we were using the wrong annotations on some of the payload fields
whose values are collections; fix those.