SW-1939 Remove unused accession summary calculation (#726)
In the older version of the web app, we used to show the total number of active
accessions in a facility or organization, but now we break the counts down by
status. Remove the code that calculated the old total. It was only referenced by
tests; the endpoint that used to call it was removed in a previous change.