SW-1939 Remove scheduled state transition dates
In v1, the user would enter dates for things like the accession being ready to
move to a drying rack, and the system would use those dates to figure out the
correct accession state at any given time.

In v2, there are no more time-based state transitions and most of the v1
user-supplied dates are no longer shown in the UI (with the exception of the
drying end date, which now just generates a notification rather than causing a
state transition).

Remove the unused date fields from `AccessionModel`.