Add missing valdation rule for negative quantities
In v1, we had a validation rule to ensure that the user-supplied total accession
quantity wasn't negative or zero.

In v2, we want to allow the user-supplied remaining quantity to be zero, but it
should never be negative. Add a validation check in the `AccessionModel`
constructor to ensure we can never have a model with a negative remaining quantity.