SW-1939 Remove v1 model validation logic
Since there are no more v1 accessions, there is no more need to validate them.

This change continues to check the `isManualState` flag to control whether or not
the v2 validation logic is executed. That's to support some unit tests that
haven't been converted to v2 yet and would fail the v2 validation checks. A
subsequent change will update those tests and remove the `isManualState` check.

Most of the tests of v1-only validation logic were removed in PR #728.