SW-1939 Remove viability test remaining quantity
In v1, we required users to enter the remaining quantity when they were
creating a viability test for a weight-based accession. Since it was user input,
we had to persist it to the database.

In v2, we require subset weight/count to be specified for weight-based accessions
in order to create viability tests, so we're always able to compute the remaining
quantity.

Drop the database columns and model field that held the v1-style value.