SW-673 Add refresh token support to test scripts
Previously, to run the test scripts, you had to manually get a valid session
cookie or access token and pass it as an argument. Add support for refresh
tokens: you can set a refresh token in an environment variable or pass it as
a command-line argument and the scripts will use it to fetch a new access
token as needed.

This is mostly useful for manual ad-hoc testing at the moment, but moves us a
little closer to being able to use these test scripts for end-to-end automated
tests.