SW-1842 Only do fuzzy matching if there are no exact matches
We want users to be able to type nonexistent accession numbers into the accession
list UI and get back close matches. But we also want users to be able to paste
exact accession numbers into the same search field and not have the results
cluttered with accessions they didn't search for.

Try to thread this needle by making the fuzzy search code start off by trying an
exact search. If that finds matches, then we're probably in the "user entered an
exact value" situation; return those results. Otherwise do the fuzzy search and
return approximate matches.

This change means there is no way for a user to do a fuzzy search for a value that
also happens to be an exact match; we believe that will still be an improvement
over the current behavior.