SW-1951 Add seedling batch CSV upload API (#762)
* SW-1951 Refactor CSV import and validation code

In preparation for adding a CSV importer for seedling batches, refactor the
species and accession CSV code to extract the common logic into superclasses.

No change in functionality here except that the error message for invalid header
rows is now consistent across accession and species CSVs.

* SW-1951 Add seedling batch CSV upload API

Add a set of endpoints under `/api/v1/nursery/batches/uploads` to import seedling
batches from CSV files.

The API is similar to the CSV upload APIs for species and accessions, but in the
case of seedling batches, there are no warning conditions that require user
confirmation. There is thus no need for endpoints to resolve problems or cancel
an upload that is awaiting confirmation.

* Remove extra slash from template path