SW-1842 Exact-then-fuzzy logic for search summaries
Commit ee97dd57 updated the search code so that if you do a fuzzy search, it first
looks for exact matches and returns those if there are any. This had the desired
effect of letting people type exact accession numbers into the accession search
UI, but the summary statistics were still based on the fuzzy matches.

Update the summary endpoint so it uses the same "try an exact search first, and
fall back to fuzzy if that doesn't return results" logic so the summary stats will
match the detailed search results.