SW-1939 Remove implicit species creation (#741)
In the v1 accessions API, clients could pass in a species name as part of the
accession data and the system would automatically add the species to the
organization's species list if it didn't already exist. In v2, clients have to
pass in the ID of an existing species.

Remove the code that supported species auto-creation.