SW-1939 Remove accession-level viability calculations (#742)
* SW-1939 Remove implicit species creation

In the v1 accessions API, clients could pass in a species name as part of the
accession data and the system would automatically add the species to the
organization's species list if it didn't already exist. In v2, clients have to
pass in the ID of an existing species.

Remove the code that supported species auto-creation.

* SW-1939 Remove accession-level viability calculations

In v1, we calculated an overall viability percentage by aggregating the results of
all the viability tests, but in v2 the viability percentage is user-selectable.
Remove the code that calculates the v1 values, and the `AccessionModel` fields
that held the v1-only values.