SW-2003 Use jOOQ JSONB for arbitrary JSON objects
In preparation for changing the default behavior of payload deserialization so
that it treats empty strings as null values, change the way we declare payload
fields that hold arbitrary client-supplied JSON objects. We will want such objects
to be preserved as-is, skipping the "treat empty strings as null" transformation,
which would be applied to `Map<String, Any?>` objects.

We could define our own class to hold arbitrary JSON objects, but since the values
will need to turn into jOOQ `JSONB` objects anyway, may as well just use that
class directly. In the payload classes, though, we'll refer to it by an alias both
to make the intent clearer and to reduce coupling between the API definition and
the database representation.

Update the user preferences API to use this new setup. This is purely an internal
change; it doesn't cause any client-visible behavior differences.