SW-1939 Remove remaining quantity from withdrawals (#748)
In v1, users were required to enter the remaining quantity when withdrawing from
weight-based accessions, to account for cases where the weights of seeds had
changed and it wouldn't be valid to just subtract the withdrawal weight from the
accession weight.

In v2, we solve that problem by letting users edit the accession remaining weight
independently of any withdrawals, so we no longer need to track remaining quantity
on a per-withdrawal basis.

Remove the database columns and model field that held the v1-style value, as well
as the backward-compatibility code that populated it based on v2-style
withdrawals.