SW-1939 Remove v1 support from AccessionModel tests (#749)
Most of the tests for `AccessionModel` use a helper method to construct a model
object. Previously, the method returned a v1 model and the tests had to explicitly
copy it to convert it to v2. Update the helper to return a v2 model, and remove
a few test methods that aren't meaningful with v2 models.