SW-1939 Remove total (inital) quantity from AccessionModel (#750)
In v1, we kept track of the quantity the user originally entered when processing
an accession, and used that as a base value for a bunch of calculations.

In v2, the user can edit the remaining quantity at any time, and we only need to
know the most recent manually-entered (observed) quantity.

There's thus no longer any need to track the initial quantity; remove it from the
data model. The corresponding database columns will be dropped in a separate
change.