SW-1939 Remove v1-only accession states from API (#753)
Stop treating v1-only accession states as valid values in the accessions API.

We can't delete the v1-only states from the `accession_states` table because
they are referenced in `accession_state_history` and those historical records
are still valid. But we no longer want them to be available as possible values
for the current state of an accession.