SW-2188 Add planting sites to search API
To support the usual filtering and sorting operations on the list of planting
sites, add them to the search API.

The `boundary` values are returned as strings whose values are JSON objects
in GeoJSON format, suitable for passing verbatim to mapping services that expect
GeoJSON values.

See `TrackingSearchTest` for an example of the field hierarchy.