SW-2161 Admin UI for importing shapefiles (#785)
Add a section to the admin UI to allow importing shapefiles to create planting
sites, planting zones, and plots. The UI allows various validation checks to be
disabled; this allows us to work with shapefiles that haven't yet been made to
comply exactly with our product requirements.

Currently this is a write-only capability. There isn't yet a way to query
existing planting site data.