SW-2105 Add CRUD endpoints for planting sites (#788)
Allow clients to read and write planting site information. Client-created planting
sites may not include planting zones or plots, and clients can't edit planting zones
or plots that were created via shapefile upload.

The planting site boundary is always returned from the GET endpoints for planting
sites created via shapefile upload. The list of planting zones and plots is always
included when querying a single planting site, but may be turned on or off when
listing the planting sites for an organization.

All the information available via the GET endpoints here is also available using
the search API, but the dedicated endpoints may be more convenient.