SW-2179 Add endpoint to get Mapbox API token (#789)
The `GET /api/v1/tracking/mapbox/token` endpoint will return a temporary Mapbox
API token that has permission to read map data.

It requires the server to be configured with a Mapbox API token that has the
`token:write` scope, e.g., in an `application.yaml` file:

```
terraware:
  mapbox:
    api-token: "YOUR_MAPBOX_API_TOKEN"
```

The implementation introduces the Ktor HTTP client library to the code base.
It is a Kotlin-native HTTP client that is a lot more convenient to use than the
Java 11 standard library HTTP client we've used previously. Existing code that
uses the Java client is left untouched here; it can be converted later.