SW-2225 Ignore fractional remaining seeds after withdrawal
When the user creates a nursery transfer or a viability test, the withdrawal
quantity is always denominated in seeds even if the accession is weight-based.
We estimate the withdrawal weight using the subset count and weight, and
subtract that estimated weight from the accession's remaining weight.

This is normally fine, but when the user is trying to withdraw all the
remaining seeds from the accession, the withdrawal ends up either failing (if
the estimated weight ended up being slightly greater than the accession's
remaining weight) or leaving the accession active even though there are no more
seeds in reality (if the estimated weight was slightly less than the
accession's remaining weight).

Add logic to handle this case: if the most recent withdrawal from a
weight-based accession is count-based, and the remaining weight would be less
than the estimated weight of one seed (positive or negative), round the
accession's remaining quantity to zero.