SW-2225 Clean up inconsistent accession quantities
Thanks to a combination of bugs in the v1 accession code and problems with the
v1->v2 migration, a handful of accessions ended up with withdrawals whose
quantities add up to more than the latest observed quantity. It's thus impossible
to do anything with the accessions because the quantity recalculation logic always
catches the problem and bails out.

Fix up the existing accessions in three ways:

* Set the correct withdrawal purpose for viability test withdrawals and add a
  database constraint to make sure it's always properly set going forward.
* Remove withdrawals that were created after the most recent quantity observation
  but don't have associated quantity history entries. These are all duplicates
  of withdrawals that _do_ have history entries, and they are causing withdrawal
  quantities to be double-counted.
* Make sure the above doesn't cause the remaining quantities (which users have
  already seen) to change suddenly, by calculating a latest observed quantity
  that causes the withdrawals to add up to the currently-displayed remaining
  quantity. (An alternative would have been to set the latest observed quantity
  to the current quantity, but that'd cause a behavior difference if a user
  edited an existing withdrawal's quantity.)

Once the cleanup has run, its class can be removed from the code base.