SW-2261 Boundaries are always of type MultiPolygon
Previously, the API schema said that the `boundary` field on planting sites,
planting zones, and plots could be of any arbitrary geometry type. But internally,
the field is constrained to always be a multipolygon. Update the API to make
that constraint clear.

Most of the change here is updating the OpenAPI schema definitions for geometry
objects so that each one only allows a single value for its `type` field.