SW-2324 Add withdrawals and deliveries to search API (#822)
Add search tables for nursery withdrawals and planting site deliveries.

The nursery withdrawals search table contains all the values needed to populate
the withdrawal log table in the UI.

Under the hood, it is implemented as a view that encapsulates the various
aggregation and concatenation operations needed to calculate the table values.
Search queries can thus sort on those derived values as if they were scalar.

The other new tables (batch withdrawals, deliveries, plantings) are ordinary
"map the underlying database table to the search API" affairs.