SW-2325 Add GET endpoints for withdrawals and deliveries (#823)
`GET /api/v1/nursery/withdrawals/{id}` returns information about a single nursery
withdrawal, including its delivery details and plantings if it is an outplanting
withdrawal. Reassignments are included in the list of plantings, as is information
about the batches the withdrawal came from (including species IDs).

`GET /api/v1/tracking/deliveries/{id}` returns information about a delivery. It
is the same delivery information that's included in the response to the withdrawals
endpoint, but without other withdrawal-related data.