SW-2361 Allow super admins to move planting sites
To support a workflow where a Terraformation employee creates a new planting site
using a shapefile, verifies it in the web app, then transfers it to the correct
organization, update the admin UI to allow moving planting sites to new
organizations.

This requires the ability to select a planting site in the admin UI; previously
it was only possible to create new sites, not view them.