Simplify OpenAPI customization
Remove the explicit list of arbitrary-JSON-object payload fields whose
`additionalProperties` values need to be cleared, in favor of detecting them
dynamically during the scan of all the payload fields we're doing anyway.

Move the GeoJSON schema configuration to a helper function that lives alongside
the schema configuration classes.

There's no change in the intended behavior here, but this does fix the definitions
of a couple of payload fields that weren't being properly adjusted previously.