SW-2003 Treat blank strings in payloads as null (#841)
Add a custom JSON deserialization module for string values that maps blank
(all-whitespace) and empty strings to null in request payloads. For example,

    {
      "a": "b",
      "c": "   "
    }

becomes equivalent to

    {
      "a": "b",
      "c": null
    }

This change means that frontend code no longer needs to check for blank form field
values; they'll be treated as missing values on the server side.

The previous behavior (treating nulls as distinct from empty strings) had an
undesirable side effect in the web app: searching for missing values using the
search API by passing in `null` as the search value would sometimes fail to return
expected results because the field in question had a non-null value that happened
to be an empty string.

Clean up all the existing blank values and add database constraints to guarantee
that we don't write any in the future.