Ignore null fuzzy search values (#843)
Commit fdd126a1 changed the behavior of the search API subtly: previously, you
could do a fuzzy search for an empty string to match all non-null values, but
now those empty strings get converted to null during payload deserialization
and the resulting searches change from matching non-null values to only matching
nulls.

The web app passes in empty-string fuzzy search values when the user hasn't
entered a value to search for. In that scenario, the previous behavior was also
wrong: we'd want to ignore the search criterion entirely and return both null
and non-null values.

Update the search implementation to ignore fuzzy search criteria if one or more
of the values is null (which, thanks to commit fdd126a1, could happen if the
client passed in an empty string).

Searching for non-null values is still possible by wrapping an exact search for
null with a `not` operation.