SW-215 Switch to maintained fork of jOOQ Gradle plugin
We're using a Gradle plugin that automatically fires up PostgreSQL in a Docker
container to run our migrations so jOOQ can look at the resulting schema for code
generation. But the plugin seems to have been abandoned; I submitted a simple pull
request over a year ago to fix a compatibility problem and it hasn't been merged
or rejected yet.

Someone forked the project and has been actively maintaining it, and their version
doesn't have the problem my PR was meant to fix. Switch to the fork and update the
build config to account for some minor naming changes the fork's maintainer has
made.