SW-1054 Support localhost HTTP redirects, round two
The hack in commit 1e987c94 failed to solve the problem in the actual staging
environment; the port number of the original request didn't make it through
the AWS load balancer plus the Nginx proxy. Try a different approach that uses
the redirect URL that's passed in by the client.