Reduce boilerplate in PermissionRequirementsTest (#847)
Nearly all the test cases in PermissionRequirementsTest check for behavior that
follows one of two patterns depending on whether they're checking a write
permission or a read permission. The only things that vary across the read
tests and across the write tests are

* The method under test
* The type of the ID of the permission's target
* The type of the exception that's thrown if the user doesn't have read access
  (this is always the same for a given ID type)
* The TerrawareUser method that controls whether or not the user has permission

Use lazy evaluation and extension methods to make a mini-DSL for these tests,
collapsing the vast majority of them to one-liners.