SW-2543 Support imperial weight units in search
Generalize the support for querying accession and accession withdrawal weights in
grams. Each grams field is now replaced with a set of weight fields, one per unit
(including grams) that does unit conversions. For example, in addition to the
existing `remainingGrams` field, we now also have `remainingKilograms`,
`remainingMilligrams`, `remainingOunces`, and `remainingPounds`.

For exact and range searches, these behave the same as the grams fields used to,
in that you can append an optional units suffix to the search terms.

But exact searches are unlikely to work like users expect in the face of unit
conversions between imperial and metric: if you search for "1 pound" you will
expect it to match an accession whose weight is entered in grams and is the
equivalent of 1.0001 pounds. So the weight field supports fuzzy searches, which
work by matching values that equal the search term if they're rounded to the
same number of decimal places as the search term. If you do a fuzzy search for
1.0 pounds, you will get back a result that weighs 1.02 pounds but not one that
weighs 1.07 pounds (since if you round that to the same number of decimal places
as the search term, it'll be 1.1, not 1.0).

Existing client code shouldn't notice this change; it can continue to do exact
and range searches on grams fields and they should return the same results as
before.