SW-1307 Improve performance of duplicate timeseries values (#870)
If an incoming timeseries values request has a lot of entries that are already in
the database, detecting the duplicates is currently pretty slow and leads to
request timeouts. Check for duplicates explicitly by querying for all of the
timestamps at once rather than relying on individual INSERT queries failing.

For a request whose values are entirely duplicates, this will result in the
server doing a single SELECT query rather than a ton of failing INSERTs.