Simplify non-nullable jOOQ queries
Use the new `asNonNullable()` extension function to remove `filterNotNull()` calls
on jOOQ queries that were only there to inform Kotlin that the results were
non-nullable. This is really just for conciseness; there will be a performance
gain from not having to copy the list, but it'll be miniscule.

In a couple places we were doing `fetch().get(FIELD)`; simplify those to
`fetchOne(FIELD)`.

No behavior change here.