WIP: Add API endpoint to get time zone names
`GET /api/v1/i18n/tz/names` (path subject to change!) returns a mapping of
time zone IDs to human-readable names in the specified locale, US English by
default. The list of time zones is based on Chrome's supported time zone list,
minus zones that appear to be duplicates.

The end result is very close to the list of time zones in the Google Calendar
time zone menu (not that Google Calendar is the ultimate authority on time
zones, just a useful comparison point).

This is missing documentation and tests at the moment; it's kind of the tail
end of a bunch of exploratory coding I was doing to figure out if it was possible
to generate this output without resorting to manual curation of the zone
information.