SW-2630 Add API endpoint to get time zone names (#881)
`GET /api/v1/i18n/timeZones` returns a list of time zone IDs and human-readable 
names in the specified locale, US English by default. The list of time zones is
based on Chrome's supported time zone list, minus zones that appear to be
duplicates.

The end result is very close to the list of time zones in the Google Calendar
time zone menu (not that Google Calendar is the ultimate authority on time
zones, just a useful comparison point).