Add test fixture for ApplicationEventPublisher
Previously, when testing classes that can publish application events, we were
using a MockK stub in place of a real `ApplicationEventPublisher` implementation.

Replace that with a dedicated test fixture class that captures the list of
published events and provides a variety of assertion methods to verify what was
published. This is less awkward to interact with than using MockK capturing slots.