SW-2637 Use seed bank time zone for withdrawal date
If the user creates a viability test and doesn't give a start date, the test is
associated with a withdrawal on today's date. Use the seed bank's time zone to
determine what "today's date" is.

This is the only place in `AccessionModel` that needs to know the time zone;
a number of methods take `Clock` arguments but they only call `Clock.instant()`.