Add test fixture for Clock
We've been using MockK to stub out `Clock` objects for testing. That works fine
when the application code just reads scalar values from the clock, but we're
shortly going to start calling `Clock.withZone()` in some places and that's
more awkward to stub out with MockK.

Replace the MockK stubs with a dedicated test class.

Most of the edits here are mechanical:

* Replace `val clock: Clock = mockk()` with `val clock = TestClock()`.
* Delete `every { clock.instant() } returns Instant.EPOCH` and
  `every { clock.zone } returns ZoneOffset.UTC` because they are the defaults.
* Replace `every { clock.instant() } returns` with `clock.instant =` for tests
  that need to change the clock on the fly.

In a few places, manual tweaking was required, e.g., passing an `Instant` to
the constructor in tests that don't use `Instant.EPOCH`.