SW-2637 Use seed bank time zone in accession logic (#893)
If the user creates a viability test and doesn't give a start date, the test is
associated with a withdrawal on today's date. Use the seed bank's time zone to
determine what "today's date" is.

Also use the seed bank's time zone when determining whether a manual quantity
adjustment falls before or after a withdrawal date.

To implement these changes, `AccessionModel` now has a required `clock`
property, which replaces the `Clock` arguments that used to need to be passed
to all the calculation methods. The new property is populated with a clock that
uses the seed bank's time zone.

Another option would have been to keep the `Clock` arguments and make sure to
always pass in a clock with the correct time zone. But that would be easier to
screw up: there'd be no good way to ensure a caller could never accidentally
pass in the system `Clock`, which would potentially cause the calculations to
return different results. With the property approach, we only need to make sure
the clock is correct in the (relatively few) places in the code where we're
constructing an `AccessionModel` from scratch, and none of the callers of the
calculation methods need to worry about it.

Unfortunately, while the application code rarely constructs `AccessionModel`s,
the same can't be said for tests. Adding a new required constructor argument
meant a lot of tests had to be touched. Since they needed to change anyway, I 
took the opportunity to introduce a new helper method to populate both the
clock and the facility ID. The result is that most of the test changes end up
removing clutter, not adding more clutter.